+++ packages/core/src/decorators/index.ts
@@ -1,1 +1,1 @@
-export { Memoized } from './memoized-getter';
+export { Memoized } from './memoized';
